spw-01 - Adjusting search tests for the sessionManager singleton
